requirements-dev.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-stubs==1.2.0
Pillow==7.1.2
python-decouple==3.3
pytz==2019.1
sqlparse==0.3.0
static3==0.7.0
